Generate resolvers in separate folder
