+++ customer-os-api/server/graph/resolver/resolver.go
@@ -1,5 +1,6 @@
-package graph
+package resolver
 
+//go:generate go run github.com/99designs/gqlgen
 // This file will not be regenerated automatically.
 //
 // It serves as dependency injection for your app, add any dependencies you require here.
